+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # vTBD
 - [BP-1552](https://movai.atlassian.net/browse/BP-1552) Search for node and subflow usage in robot
+  - Add mobdata seatch commands to search for Node and Flow usage
 
 # v3.16.2
 - [BP-1420](https://movai.atlassian.net/browse/BP-1420): Skip removal of Package:maps and Package:point_clouds on import